Compatibility with redis 4.7.x gem
It's not clear why this complexity was added in https://github.com/rpush/modis/pull/32 but it was unnecessary and breaks the test suite with redis 4.7.1. Removing it continues with tests passing on 4.6.0 and fixes the tests with 4.7.1.

Compatibility with redis 4.8.x will come in https://github.com/rpush/modis/pull/35.

See https://github.com/rpush/modis/pull/36.